+++ packages/connect-wallet-modal/src/components/WalletsModalForEth/WalletsModalForEth.tsx
@@ -1,9 +1,6 @@
 import React from 'react';
 import { helpers } from '@reef-knot/web3-react';
 import { useReefKnotContext } from '@reef-knot/core-react';
-import { WalletConnectConnector } from 'wagmi/connectors/walletConnect';
-import { WalletConnectLegacyConnector } from 'wagmi/connectors/walletConnectLegacy';
-import { InjectedConnector } from 'wagmi/connectors/injected';
 import { WalletAdapterData } from '@reef-knot/types';
 import {
   ConnectBraveWallet,
@@ -25,8 +22,8 @@ import { WalletsModalForEthProps } from './types';
 import { WALLET_IDS, WalletId } from '../../constants';
 
 const walletsButtons: { [K in WalletId | string]: React.ComponentType } = {
-  Injected: ConnectInjected,
-  WalletConnect: ConnectWC,
+  injected: ConnectInjected,
+  walletConnect: ConnectWC,
   [WALLET_IDS.METAMASK]: ConnectMetamask,
   [WALLET_IDS.LEDGER]: ConnectLedger,
   [WALLET_IDS.COINBASE]: ConnectCoinbase,
@@ -104,18 +101,7 @@ function getWalletsButtons(
       (data) => data.walletId === walletId,
     );
     if (walletData) {
-      let connectorId = '';
-      // Multiple wagmi version can cause problems here (as objects will not directly equal)
-      // TODO: rework to a better solution
-      if (
-        walletData.connector instanceof WalletConnectLegacyConnector ||
-        walletData.connector instanceof WalletConnectConnector
-      ) {
-        connectorId = 'WalletConnect';
-      }
-      if (walletData.connector instanceof InjectedConnector) {
-        connectorId = 'Injected';
-      }
+      const connectorId = walletData.connector.id;
       return getWalletButton(connectorId, walletId, {
         ...commonProps,
         ...walletData,