+++ packages/wallets-helpers/src/factories/index.ts
@@ -1,2 +1,1 @@
 export * from './walletConnectConnector';
-export * from './walletConnectMobileLinks';